+++ sim/state/bitmap.ts
@@ -1105,3 +1105,27 @@ namespace pxsim.bitmap {
         return toBuffer(img)
     }
 }
+
+namespace pxsim.pxtcore {
+    export function updateScreen(img: RefImage) {
+
+    }
+    export function updateStats(s: string) {
+
+    }
+    export function setPalette(b: RefBuffer) {
+
+    }
+    export function setScreenBrightness(b: number) {
+   
+    }
+    export function displayHeight(): number {
+        return 120
+    }
+    export function displayWidth(): number {
+        return 160
+    }
+    export function displayPresent(): boolean {
+        return true
+    }
+}
